api/pyproject.toml
[project]
name = "bidforge-api"
version = "0.1.0"
description = "BidForge FastAPI orchestration — Langfuse, Supabase, agent pipeline"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.32",
  "pydantic-settings>=2.6",
  "pyjwt[crypto]>=2.9",
  "cryptography>=43",
  "httpx>=0.27",
  "openai>=1.59",
  "langfuse>=2.57",
  "supabase>=2.10",
  "bidforge-schemas",
  "bidforge-prompts",
  "bidforge-shared",
  "bidforge-agents",
  "fpdf2>=2.8",
  "pypdf>=5.0",
  "python-docx>=1.1",
  "python-multipart>=0.0.9",
]

[dependency-groups]
dev = ["ruff>=0.8", "pytest>=8.3", "httpx>=0.27"]

[tool.uv.sources]
bidforge-schemas = { path = "../packages/schemas", editable = true }
bidforge-prompts = { path = "../packages/prompts", editable = true }
bidforge-shared = { path = "../packages/shared", editable = true }
bidforge-agents = { path = "../packages/agents", editable = true }

[tool.uv]
package = false

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
markers = [
    "integration: calls live OpenRouter (requires OPENROUTER_API_KEY)",
]
